+++ client/src/pages/Community/PostDetail.tsx
@@ -11,13 +11,13 @@ import Swal from 'sweetalert2';
 
 import { Viewer } from '@toast-ui/react-editor';
 
-import Answers from '../../Components/ccc/Answers';
-import ReviewCarousel from '../../Components/ccc/ReviewCarousel';
+import Answers from '../../Components/community/Answers';
+import ReviewCarousel from '../../Components/community/ReviewCarousel';
 import { Api } from '../../apis/customAPI';
 import useAxios from '../../hooks/useAxios';
 import { Iuser } from '../../reducers/userInfoReducer';
 import { RootState } from '../../store/Store';
-import TopBar from '../../Components/ccc/TopBar';
+import TopBar from '../../Components/community/TopBar';
 
 const TopBarContainer = styled.div`
 	width: 90%;
